requirements.txt
loguru==0.7.2
PyYAML==6.0.1
